+++ src/com/neurix/simrs/master/ruangan/model/MtSimrsRuanganEntity.java
@@ -1,7 +1,6 @@
 package com.neurix.simrs.master.ruangan.model;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.sql.Timestamp;
 
 public class MtSimrsRuanganEntity implements Serializable {
@@ -12,7 +11,7 @@ public class MtSimrsRuanganEntity implements Serializable {
     private String statusRuangan;
     private String idKelasRuangan;
     private String keterangan;
-    private BigInteger tarif;
+    private Long tarif;
     private String branchId;
     private String flag;
     private String action;
@@ -69,11 +68,11 @@ public class MtSimrsRuanganEntity implements Serializable {
         this.keterangan = keterangan;
     }
 
-    public BigInteger getTarif() {
+    public Long getTarif() {
         return tarif;
     }
 
-    public void setTarif(BigInteger tarif) {
+    public void setTarif(Long tarif) {
         this.tarif = tarif;
     }
